reoganized code
