+++ src/modules/handleEvent.js
@@ -1,0 +1,58 @@
+import TaskManager from './renderPage.js';
+import refreshImg from '../assets/refresh.svg';
+import deleteImg from '../assets/delete.svg';
+
+const taskManager = new TaskManager('#task-list');
+const refresh = document.getElementById('refreshSvg');
+const menu = document.getElementById('menu');
+let editInput = false;
+
+refresh.src = refreshImg;
+
+const handleFormSubmission = (e) => {
+  e.preventDefault();
+  const target = e.target;
+
+  if (target.matches('#todoForm')) {
+    const input = document.querySelector('#enter-description');
+    const description = input.value.trim();
+
+    if (description !== '') {
+      taskManager.addTask(description);
+      input.value = '';
+    }
+  } else if (target.matches('#todo-edit')) {
+    const input = document.querySelector('#edit-input');
+    taskManager.updateTaskDescription(input.value.trim());
+  }
+}
+
+const handleClicks = (e) => {
+  const target = e.target;
+
+  if (target.matches('#clearCompleted')) {
+    taskManager.clearCompleted();
+  } else if (target.matches('.completed-checkbox')) {
+    taskManager.updateTaskCompletion(target);
+  } else if (target.matches('.treeDots')) {
+    const index = [...document.querySelectorAll('.treeDots')].indexOf(target);
+    target.src = deleteImg;
+    target.id = 'treeDots-delete';
+    taskManager.getIndex(index);
+    editInput = true;
+  } else if (editInput && target.matches('*:not(#edit-input)')) {
+    taskManager.renderTasks();
+    editInput = false;
+  } else if (target.matches('.delete-task')) {
+    taskManager.deleteTask();
+    while (menu.firstChild) {
+      menu.firstChild.remove();
+    }
+  } else if (target.matches('.task-description')) {
+    editInput = true;
+    const index = [...document.querySelectorAll('.task-description')].indexOf(target);
+    target.innerHTML = taskManager.editDescription(index);
+  }
+}
+
+export { handleFormSubmission, handleClicks }
