+++ webpack.config.js
@@ -8,7 +8,7 @@ module.exports = {
   },
   output: {
     filename: '[name][contenthash].js',
-    path: path.resolve(__dirname, 'dist'),
+    path: path.resolve(__dirname, 'docs'),
     clean: true,
   },
   plugins: [
@@ -20,7 +20,7 @@ module.exports = {
   devtool: 'source-map',
   devServer: {
     static: {
-      directory: path.resolve(__dirname, 'dist'),
+      directory: path.resolve(__dirname, 'docs'),
     },
     port: 3000,
     open: true,
